requirements.txt
termcolor
requests
pyOpenSSL
ndg-httpsclient
pyasn1
feedparser
bs4
stanfordcorenlp
pony
pymysql
